Seperated the FileTree section from the INputPage class.
